+++ packages/flutter_riverpod/lib/src/builders.dart
@@ -8,6 +8,8 @@
 // You can then use it in your terminal by executing:
 // generate_providers <riverpod/flutter_riverpod/hooks_riverpod> <path to builder file to update>
 
+// ignore_for_file: invalid_use_of_internal_member
+
 import 'package:flutter/foundation.dart';
 import 'package:meta/meta.dart';
 
@@ -231,11 +233,13 @@ class ChangeNotifierProviderFamilyBuilder {
     NotifierT Function(Ref<NotifierT> ref, ArgT param) create, {
     String? name,
     Iterable<ProviderOrFamily>? dependencies,
+    Retry? retry,
   }) {
     return ChangeNotifierProviderFamily<NotifierT, ArgT>(
       create,
       name: name,
       dependencies: dependencies,
+      retry: retry,
     );
   }
 
@@ -301,12 +305,14 @@ class AutoDisposeChangeNotifierProviderBuilder {
     NotifierT Function(Ref<NotifierT> ref) create, {
     String? name,
     Iterable<ProviderOrFamily>? dependencies,
+    Retry? retry,
   }) {
     return ChangeNotifierProvider<NotifierT>(
       create,
       name: name,
       isAutoDispose: true,
       dependencies: dependencies,
+      retry: retry,
     );
   }
 
@@ -325,12 +331,14 @@ class AutoDisposeChangeNotifierProviderFamilyBuilder {
     NotifierT Function(Ref<NotifierT> ref, ArgT param) create, {
     String? name,
     Iterable<ProviderOrFamily>? dependencies,
+    Retry? retry,
   }) {
     return ChangeNotifierProviderFamily<NotifierT, ArgT>(
       create,
       name: name,
       isAutoDispose: true,
       dependencies: dependencies,
+      retry: retry,
     );
   }
 }