+++ packages/riverpod/lib/src/providers/provider.dart
@@ -22,6 +22,7 @@ base class Provider<StateT> extends $FunctionalProvider<StateT, StateT>
     super.name,
     super.dependencies,
     super.isAutoDispose = false,
+    super.retry,
   }) : super(
           allTransitiveDependencies:
               computeAllTransitiveDependencies(dependencies),
@@ -39,6 +40,7 @@ base class Provider<StateT> extends $FunctionalProvider<StateT, StateT>
     required super.from,
     required super.argument,
     required super.isAutoDispose,
+    required super.retry,
   });
 
   /// {@macro riverpod.autoDispose}
@@ -69,6 +71,7 @@ base class Provider<StateT> extends $FunctionalProvider<StateT, StateT>
       from: from,
       argument: argument,
       isAutoDispose: isAutoDispose,
+      retry: retry,
       allTransitiveDependencies: null,
       dependencies: null,
       name: null,
@@ -366,6 +369,7 @@ class ProviderFamily<StateT, ArgT>
     super.name,
     super.dependencies,
     super.isAutoDispose = false,
+    super.retry,
   }) : super(
           providerFactory: Provider.internal,
           allTransitiveDependencies:
@@ -380,5 +384,6 @@ class ProviderFamily<StateT, ArgT>
     required super.dependencies,
     required super.allTransitiveDependencies,
     required super.isAutoDispose,
+    required super.retry,
   }) : super(providerFactory: Provider.internal);
 }